% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_nested_data}
\alias{example_nested_data}
\title{Example nested data}
\format{
A tibble with 100 rows and 4 variables
\describe{
\item{id}{the ID, either 1 or 2}
\item{x, y, z}{the columns to nest}
}
}
\usage{
example_nested_data
}
\description{
A dataset containing example data that can be nested. Mainly used for
examples and testing.
}
\examples{

example_nested_data
}
\keyword{datasets}
