% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_nested_tune_results}
\alias{example_nested_tune_results}
\title{Example nested tuning results}
\format{
A tibble with 4 rows and 5 columns, and additional classes
"nested_tune_results" and "tune_results".
\describe{
\item{nest_id}{The nested data frame that the tune results correspond to}
\item{splits, id, .metrics, .notes}{Normal columns returned by \code{tune_grid()}}
}
}
\usage{
example_nested_tune_results
}
\description{
The results of using \code{\link[=tune_nested]{tune_nested()}} on example nested data. Specifically,
the results of tuning a 'nested_workflows' object: a set of workflows
corresponding to each nested data frame, each tuned individually on said
data frame.
}
\details{
Generated by the following code:

\if{html}{\out{<div class="sourceCode">}}\preformatted{set.seed(42)

model <- parsnip::linear_reg(penalty = tune()) \%>\%
  parsnip::set_engine("glmnet") \%>\%
  nested()

recipe <- recipes::recipe(example_nested_data, z ~ x + y + id) \%>\%
  recipes::step_pca(x, y, num_comp = tune::tune()) \%>\%
  step_nest(-id)

wf <- workflows::workflow() \%>\%
  workflows::add_recipe(recipe) \%>\%
  add_nested_model(model)

wfs <- list(wf, wf) \%>\%
  nested()

resamples <- nested_resamples(example_nested_data,
                              rsample::vfold_cv, recipe, 2)

example_nested_tune_results <- tune_nested(
  wfs, tune::tune_grid, resamples
)
}\if{html}{\out{</div>}}
}
\examples{

example_nested_tune_results
}
\keyword{datasets}
