% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nested_split.R
\name{extract_nested_split}
\alias{extract_nested_split}
\title{Extract a single split from more than one nested split.}
\usage{
extract_nested_split(x, ...)
}
\arguments{
\item{x}{A 'nested_resamples' object.}

\item{n}{A number. The nth resample set will be extracted (negative numbers
are accepted, and extract the nth set from the end.)}
}
\value{
A tibble of resamples.
}
\description{
'extract_nested_split' allows you to extract a single nested split from
multiple.
}
\details{
If \code{x} is not a 'nested_resamples' object, 'extract_nested_resamples' will
return the nth object in the 'splits' column.
This function is designed to format nested resamples for
\code{\link[=extract_nested_training]{extract_nested_training()}} and similar.
As a consequence, the result of this function is not actually a single split
object, but a tibble of resamples. However, each split corresponds to a
specific nest in your data, so using \code{\link[=extract_nested_training]{extract_nested_training()}} (or
similar) will return a single tibble.
Since it is possible for resamples in different nests to be different sizes,
if the \code{n}th set does not exist for some of the nests in \code{x}, only valid
resamples will be returned.
}
\seealso{
\code{\link[=extract_nested_training]{extract_nested_training()}} and \code{\link[=extract_nested_testing]{extract_nested_testing()}} for
extracting elements of a single nested split.
}
