% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nested_resamples.R
\name{extract_nested_training}
\alias{extract_nested_training}
\title{Nested Training/Testing Set Splitting}
\usage{
extract_nested_training(x)
}
\arguments{
\item{x}{A 'nested_resamples' object produced by \code{\link[=nested_resamples]{nested_resamples()}}.}
}
\value{
A data frame corresponding to the testing/training split.
}
\description{
These functions allow the extraction of the training and testing data from a
split.
}
\details{
The number of rows of \code{x} must be the same as the number of nests in the
data.
This means that \code{x} must be a single nested split.
This should most commonly be used on nested resamples with \code{.f} being
\code{\link[rsample:initial_split]{rsample::initial_split()}}, or similar.
If \code{x} is not a 'nested_resamples' object, it will be passed onto
\code{\link[rsample:initial_split]{rsample::testing()}} / \code{\link[rsample:initial_split]{rsample::training()}}.
}
\seealso{
\code{\link[=extract_nested_split]{extract_nested_split()}} for extracting a single nested split
from multiple splits.
}
