% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_resamples.R
\name{fit_resamples.nested_model}
\alias{fit_resamples.nested_model}
\alias{fit_resamples.nested_models}
\alias{fit_resamples.nested_workflow}
\alias{fit_resamples.nested_workflows}
\title{Fit multiple nested models via resampling}
\usage{
\method{fit_resamples}{nested_model}(
  object,
  preprocessor,
  resamples,
  ...,
  metrics = NULL,
  control = control_nested()
)

\method{fit_resamples}{nested_models}(
  object,
  preprocessor,
  resamples,
  ...,
  metrics = NULL,
  control = control_nested()
)

\method{fit_resamples}{nested_workflow}(
  object,
  resamples,
  ...,
  metrics = NULL,
  control = control_nested()
)

\method{fit_resamples}{nested_workflows}(
  object,
  resamples,
  ...,
  metrics = NULL,
  control = control_nested()
)
}
\arguments{
\item{object}{A 'nested_model', 'nested_models', 'nested_workflow' or
'nested_workflows' object.}

\item{preprocessor}{A model formula, recipe, 'nested_formulas' or
'nested_recipes' object.}

\item{resamples}{A 'nested_rset' object created using \code{\link[=nested_resamples]{nested_resamples()}}.}

\item{...}{Passed directly into \code{\link[tune:fit_resamples]{tune::fit_resamples()}}. Currently unused.}

\item{metrics}{A \code{\link[yardstick:metric_set]{yardstick::metric_set()}} or NULL. Passed directly into
\code{\link[tune:fit_resamples]{tune::fit_resamples()}}.}

\item{control}{A \code{\link[tune:control_grid]{tune::control_resamples()}} or \code{\link[=control_nested]{control_nested()}} object.
Use \code{\link[=control_nested]{control_nested()}} if you want to customize the fitting process over the
nested data frames and resamples individually.}
}
\value{
A tibble of resample results
}
\description{
\code{fit_resamples} methods for nested models or workflows. See
\code{\link[tune:fit_resamples]{tune::fit_resamples()}}.
}
\examples{
data("example_nested_data")

model <- model <- parsnip::svm_rbf(mode = "regression", cost = tune()) \%>\%
  parsnip::set_engine("kernlab") \%>\%
  nested()

recipe <- recipes::recipe(example_nested_data, z ~ x + y + id) \%>\%
  recipes::step_pca(x, y, num_comp = tune::tune()) \%>\%
  step_nest(id)

workflow <- workflows::workflow() \%>\%
  workflows::add_recipe(recipe) \%>\%
  add_nested_model(model)

resamples <- nested_resamples(recipe, rsample::vfold_cv, data, 2)

fit_resamples(workflow, resamples)

}
\seealso{
\code{\link[tune:fit_resamples]{tune::fit_resamples()}}
}
