% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_initial_split.R
\name{nested_initial_split}
\alias{nested_initial_split}
\alias{nested_initial_time_split}
\alias{nested_group_initial_split}
\title{Nested Training/Test Set Splitting}
\usage{
nested_initial_split(
  data,
  prop = 3/4,
  strata = NULL,
  breaks = 4,
  pool = 0.1,
  nesting_method = NULL,
  ...
)

nested_initial_time_split(
  data,
  prop = 3/4,
  lag = 0,
  nesting_method = NULL,
  ...
)

nested_group_initial_split(data, group, prop = 3/4, nesting_method = NULL, ...)
}
\arguments{
\item{data}{A data frame.}

\item{prop}{Passed into \code{\link[rsample:initial_split]{rsample::initial_split()}} or
\code{\link[rsample:initial_split]{rsample::initial_time_split()}}.}

\item{strata}{Passed into \code{\link[rsample:initial_split]{rsample::initial_split()}}.}

\item{breaks}{Passed into \code{\link[rsample:initial_split]{rsample::initial_split()}}.}

\item{pool}{Passed into \code{\link[rsample:initial_split]{rsample::initial_split()}}.}

\item{nesting_method}{A recipe, workflow or NULL.}

\item{...}{Passed into the respective rsample function.}

\item{lag}{Passed into \code{\link[rsample:initial_split]{rsample::initial_time_split()}}.}

\item{group}{Passed into \code{\link[rsample:initial_split]{rsample::group_initial_split()}}.}
}
\description{
These functions allow you to split nested data into training and testing
data.
}
\details{
These functions take nested data, and perform sampling within each nest. This
normally ensures that data from every nest will be included
}
