% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedmodels-package.R
\docType{package}
\name{nestedmodels-package}
\alias{nestedmodels}
\alias{nestedmodels-package}
\title{nestedmodels: Tidy Modelling for Nested/Panel Data}
\description{
The goal of 'nestedmodels' is to allow the modelling of nested data. Some models only accept certain predictors. For panel data, it is often desirable to create a model for each panel. 'nestedmodels' enhances the 'tidymodels' set of packages by allowing the user to classify a model or workflow as 'nested'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ashbythorpe/nestedmodels}
  \item \url{https://ashbythorpe.github.io/nestedmodels/}
  \item Report bugs at \url{https://github.com/ashbythorpe/nestedmodels/issues}
}

}
\author{
\strong{Maintainer}: Ashby Thorpe \email{ashbythorpe@gmail.com} (\href{https://orcid.org/0000-0003-3106-099X}{ORCID})

}
\keyword{internal}
