% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.nested_model_fit}
\alias{tidy.nested_model_fit}
\alias{tidy.nested_workflow_fit}
\alias{augment.nested_model_fit}
\alias{augment.nested_workflow_fit}
\alias{glance.nested_model_fit}
\alias{glance.nested_workflow_fit}
\alias{tidy.nested_resamples}
\alias{augment.nested_tune_results}
\alias{augment.nested_resample_results}
\alias{augment.nested_last_fit}
\title{Tidy, augment and glance methods for nested objects}
\usage{
\method{tidy}{nested_model_fit}(x, ...)

\method{tidy}{nested_workflow_fit}(x, ...)

\method{augment}{nested_model_fit}(object, new_data, ...)

\method{augment}{nested_workflow_fit}(object, new_data, ...)

\method{glance}{nested_model_fit}(x, ...)

\method{glance}{nested_workflow_fit}(x, ...)

\method{tidy}{nested_resamples}(x, ...)

\method{augment}{nested_tune_results}(x, parameters = NULL, ...)

\method{augment}{nested_resample_results}(x, ...)

\method{augment}{nested_last_fit}(x, ...)
}
\arguments{
\item{x}{The object to be converted into a tidy \code{\link[tibble:tibble]{tibble::tibble()}}.}

\item{...}{Additional arguments passed into their respective functions.
(e.g. for \code{tidy.nested_rset()}, \code{\link[rsample:tidy.rsplit]{rsample::tidy.rset()}}).}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}. For \code{glance()}, the tibble will have 1 row.
}
\description{
These methods allow the usage of \code{broom} functions on nested objects.
}
\details{
Tidying a nested object generally involves applying an already defined
\code{broom} method to a set of nested data frames or a set of objects produced
from nested data frames, and then combining the result back into a tibble.
}
\section{Glance}{

\code{\link[generics:glance]{generics::glance()}} states that \code{glance} methods should always return 1
row outputs for non-empty inputs. The 'nestedmodels' package is no
exception: \code{glance} methods will combine rows to produce a result with a
single row. Specifically:
\itemize{
\item If a column contains 1 unique value, that value is used.
\item If a column is numeric, the mean will be calculated.
\item Otherwise, the results will be combined into a list
}

This may not give you the most useful of results.
If you want to generate non-combined glances about every nested data frame,
use:
\code{purrr::map(x, glance)}, \code{lapply(x, glance)}
or similar.
}

\examples{
data("example_nested_data")
resamples <- nested_resamples(id, rsample::vfold_cv, example_nested_data)
tidy(resamples)

}
