% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_nested.R
\name{tune_nested}
\alias{tune_nested}
\alias{tune_nested.nested_model}
\alias{tune_nested.nested_models}
\alias{tune_nested.nested_workflow}
\alias{tune_nested.nested_workflows}
\title{Nested model tuning}
\usage{
tune_nested(object, ...)

\method{tune_nested}{nested_model}(
  object,
  .f,
  preprocessor,
  resamples,
  ...,
  control = control_nested()
)

\method{tune_nested}{nested_models}(
  object,
  .f,
  preprocessor,
  resamples,
  ...,
  control = control_nested()
)

\method{tune_nested}{nested_workflow}(object, .f, resamples, ..., control = control_nested())

\method{tune_nested}{nested_workflows}(object, .f, resamples, ..., control = control_nested())
}
\arguments{
\item{object}{An object with class 'nested_model', 'nested_workflow',
'nested_models' and 'nested_workflows'.}

\item{...}{Passed into \code{.f}.}

\item{.f}{A function, function name, or purrr style anonymous function to
tune models (see \code{\link[rlang:as_function]{rlang::as_function()}}).}

\item{preprocessor}{A traditional model formula or a recipe created using
\code{\link[recipes:recipe]{recipes::recipe()}}.}

\item{resamples}{A 'nested_rset' object generated by \code{\link[=nested_resamples]{nested_resamples()}}.}

\item{control}{An object created using \code{\link[=control_nested]{control_nested()}} or the control
object corresponding to \code{.f} (e.g. \code{\link[tune:control_grid]{tune::control_grid()}})}
}
\value{
A tibble of tune results.
}
\description{
\code{tune_nested} allows you to use any tuning function on nested models or
workflows.
}
\details{
\code{.f} (or a method of \code{.f}) must take a workflow as its first argument and a
resample object as its second. Normally, \code{tune_nest} will combine the
resamples and the model/workflow will be tuned normally. However, if you
pass in multiple models/workflows, the resample sets will be used
individually on each model/workflow. The resulting tables will be combined
by row binding and adding a column named 'nest_id', corresponding to the
resample set that the model/workflow was tuned on.
}
\seealso{
\code{\link[tune:tune_grid]{tune::tune_grid()}} \code{\link[tune:tune_bayes]{tune::tune_bayes()}}
}
