% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedmodels-package.R
\docType{package}
\name{nestedmodels-package}
\alias{nestedmodels}
\alias{nestedmodels-package}
\title{nestedmodels: Tidy Modelling for Nested Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A modelling framework for nested data using the 'tidymodels' ecosystem. Specify how to nest data using the 'recipes' package, create testing and training splits using 'rsample', and fit models to this data using the 'parsnip' and 'workflows' packages. Allows any model to be fit to nested data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ashbythorpe/nestedmodels}
  \item \url{https://ashbythorpe.github.io/nestedmodels/}
  \item Report bugs at \url{https://github.com/ashbythorpe/nestedmodels/issues}
}

}
\author{
\strong{Maintainer}: Ashby Thorpe \email{ashbythorpe@gmail.com} (\href{https://orcid.org/0000-0003-3106-099X}{ORCID})

}
\keyword{internal}
